Makes the implementation generic
